backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
websockets>=11.0.3
groq>=0.4.0
google-generativeai>=0.3.0
